Collabora: hide loading delay warning if document is loaded
