+++ scripts/analysis/findbugs-results.txt
@@ -1,1 +1,1 @@
-473
+474
